### 2025-04-30 bb93ade
Update paper.bib
removed https://doi.org prefixes in doi's
