+++ paper/paper.bib
@@ -5,7 +5,7 @@
   pages = {1-17},
   year = {1986},
   issn = {0003-2670},
-  doi = {https://doi.org/10.1016/0003-2670(86)80028-9},
+  doi = {10.1016/0003-2670(86)80028-9},
   url = {https://www.sciencedirect.com/science/article/pii/0003267086800289},
   author = {Paul Geladi and Bruce R. Kowalski}
 }
@@ -43,7 +43,7 @@
   pages = {360 -- 378},
   year = {1931},
   doi = {10.1214/aoms/1177732979},
-  URL = {https://doi.org/10.1214/aoms/1177732979}
+  URL = {10.1214/aoms/1177732979}
 }
 
 
@@ -55,7 +55,7 @@
   pages = {45-65},
   year = {1996},
   issn = {0169-7439},
-  doi = {https://doi.org/10.1016/S0169-7439(96)00007-X},
+  doi = {10.1016/S0169-7439(96)00007-X},
   url = {https://www.sciencedirect.com/science/article/pii/S016974399600007X},
   author = {Philip R.C. Nelson and Paul A. Taylor and John F. MacGregor},
   keywords = {PCA, PLS, Missing data, NIPALS algorithm, EM algorithm},
@@ -70,7 +70,7 @@
   number = {1},
   pages = {93-109},
   keywords = {batch process monitoring, abnormal situation detection, trajectory alignment, transition control, grade-to-grade transitions, process start-up, multiway PCA, multiway PLS, multiblock multiway PCA/PLS},
-  doi = {https://doi.org/10.1002/cem.778},
+  doi = {10.1002/cem.778},
   url = {https://analyticalsciencejournals.onlinelibrary.wiley.com/doi/abs/10.1002/cem.778},
   eprint = {https://analyticalsciencejournals.onlinelibrary.wiley.com/doi/pdf/10.1002/cem.778},
   abstract = {Abstract There has been a lot of research activity in the area of batch process analysis and monitoring for abnormal situation detection since the pioneer work of Nomikos and MacGregor [1–5]. However, some of the key ideas and the thought process that led to those first papers have been forgotten. Batch process data are dynamic data. The whole philosophy of looking at batch process data with latent variables was developed because batch process variables are both autocorrelated and cross-correlated. Statistical process control by definition checks deviations from a nominal behavior (a target). Therefore for statistical process control of batch processes we should look at deviations of process variable trajectories from their nominal trajectories and from their nominal auto/cross-correlations. An added advantage to modeling the deviations from the target trajectory is that a non-linear problem is converted to a linear one that it is easy to tackle with linear latent variable methods such as principal component analysis (PCA) and partial least squares (PLS). This paper first takes a critical look at the true nature of batch process data. The general case where variables are not present during the entire duration of the batch is addressed. It is then illustrated how proper centering (by taking the deviations from the target trajectory) can retain valuable information on auto- and cross-correlation of the process variables. This auto- and cross-correlation is only modeled with a certain types of models. Topics such as scaling and trajectory alignment are revisited and issues arising when using the indicator variable approach are addressed. The development of control charts for multiblock, multiway PCA/PLS is discussed. Practical issues related to applications in industry are addressed. Then some of the methods that have appeared in the literature are examined as to their assumptions, their advantages and disadvantages and their range of applicability. Finally the nature of transition data (start-ups, grade transitions) is discussed and issues related to aligning, centering and scaling such types of data are presented. Copyright © 2003 John Wiley \& Sons, Ltd.},
@@ -104,7 +104,7 @@
   volume = {30},
   number = {1},
[…]
