### 2025-05-22 3c2acd6
Enable using explicit arguments like `input_scores=`
- invoking base classes _BasePCA and _PLS creates mess with our implementation, since they specify required input arguments, too
- removed base classes from nipalsPCA and nipalsPLS
- made arguments explicit in unit tests
- removed some remaining cases of internal variables containing capital letters in nipalsPLS.py
