+++ src/open_nipals/nipalsPLS.py
@@ -24,13 +24,12 @@ from __future__ import (
     annotations,
 )  # needed so we can return NipalsPLS class in our type hints
 import numpy as np
-from sklearn.cross_decomposition._pls import _PLS
 import warnings
 from open_nipals.utils import _nan_mult
 from typing import Optional, Tuple, Union
 
 
-class NipalsPLS(_PLS):
+class NipalsPLS:
     def __init__(
         self,
         n_components: int = 2,
@@ -63,11 +62,11 @@ class NipalsPLS(_PLS):
         self.fit_data_x = None  # n_rows_x x n_cols_x
         self.fit_data_y = None  # n_rows_y x n_cols_y
 
-        self.loadings_X = None  # n_cols_x x n_components
+        self.loadings_x = None  # n_cols_x x n_components
         self.fit_scores_x = None  # n_rows_x x n_components
-        self.weights_X = None  # n_cols_x x n_components
+        self.weights_x = None  # n_cols_x x n_components
 
-        self.loadings_Y = None  # n_cols_y x n_components
+        self.loadings_y = None  # n_cols_y x n_components
         self.fit_scores_y = None  # n_rows_y x n_components
 
         self.regression_matrix = None  # n_components x n_predictors
@@ -85,8 +84,8 @@ class NipalsPLS(_PLS):
         Returns:
             int
         """
-        if self.loadings_X is not None:
-            return self.loadings_X.shape[1]
+        if self.loadings_x is not None:
+            return self.loadings_x.shape[1]
         else:
             return 0
 
@@ -163,7 +162,7 @@ class NipalsPLS(_PLS):
             input_y = input_y - sim_data_y
 
             p = np.append(
-                self.loadings_X, np.zeros((n_cols_x, n_add)), axis=1
+                self.loadings_x, np.zeros((n_cols_x, n_add)), axis=1
             )  # x loadings
             t = np.append(
                 self.fit_scores_x, np.zeros((n_rows_x, n_add)), axis=1
@@ -172,10 +171,10 @@ class NipalsPLS(_PLS):
                 self.fit_scores_y, np.zeros((n_rows_x, n_add)), axis=1
             )  # y scores
             w = np.append(
-                self.weights_X, np.zeros((n_cols_x, n_add)), axis=1
+                self.weights_x, np.zeros((n_cols_x, n_add)), axis=1
             )  # weights
             q = np.append(
-                self.loadings_Y, np.zeros((n_cols_y, n_add)), axis=1
+                self.loadings_y, np.zeros((n_cols_y, n_add)), axis=1
             )  # y loading
             b = np.zeros(
                 (fitted_components + n_add, fitted_components + n_add)
@@ -289,10 +288,10 @@ class NipalsPLS(_PLS):
             # End of for loop
 
         self.fit_scores_x = t
-        self.loadings_X = p
-        self.weights_X = w
+        self.loadings_x = p
+        self.weights_x = w
         self.fit_scores_y = u
-        self.loadings_Y = q
+        self.loadings_y = q
         self.regression_matrix = b
 
     def set_components(self, n_component: int, verbose: bool = False):
@@ -393,22 +392,22 @@ class NipalsPLS(_PLS):
         """
 
         # Check whether the model is available or not
-        if self.loadings_X is None:
+        if self.loadings_x is None:
             raise ValueError("Model has not yet been fit.")
 
         if input_y is None:
-            scores_x = self._transform_XY(
-                input_x, self.loadings_X, weights=self.weights_X
+            scores_x = self._transform_xy(
+                input_x, self.loadings_x, weights=self.weights_x
             )
             return scores_x
         else:
-            scores_x = self._transform_XY(
-                input_x, self.loadings_X, weights=self.weights_X
[…]
