+++ tests/test_nipalsPLS_RandomGen.py
@@ -108,7 +108,7 @@ def fitted_model_pass_dat(
     scaler_x, dat_x_scaled = init_scaler(x)
     scaler_y, dat_y_scaled = init_scaler(y)
     pls_model = NipalsPLS(mean_centered=True)  # pylint: disable=not-callable
-    pls_model.fit(dat_x_scaled, dat_y_scaled)
+    pls_model.fit(input_x=dat_x_scaled, input_y=dat_y_scaled)
     return pls_model, scaler_x, scaler_y
 
 
@@ -222,9 +222,13 @@ class TestFit(unittest.TestCase):
         if "NaN" in self.name:
             with warnings.catch_warnings():
                 warnings.filterwarnings("ignore")
-                py_calc_scores = model.transform(scaler_x.transform(in_data))
+                py_calc_scores = model.transform(
+                    input_x=scaler_x.transform(in_data)
+                )
         else:
-            py_calc_scores = model.transform(scaler_x.transform(in_data))
+            py_calc_scores = model.transform(
+                input_x=scaler_x.transform(in_data)
+            )
 
         test_val = rmse(model.fit_scores_x, py_calc_scores)
         lin_val = nan_conc_coeff(model.fit_scores_x, py_calc_scores)
@@ -238,8 +242,8 @@ class TestFit(unittest.TestCase):
 
     def test_fit_loadings(self):
         """Compare loadings to loadings from package (P)"""
-        test_val = rmse(self.P, self.model[0].loadings_X)
-        lin_val = nan_conc_coeff(self.P, self.model[0].loadings_X)
+        test_val = rmse(self.P, self.model[0].loadings_x)
+        lin_val = nan_conc_coeff(self.P, self.model[0].loadings_x)
         with self.subTest():
             # overall rmse is low
             self.assertLess(test_val, 1e-3, msg=f"rmse = {test_val}")
@@ -318,7 +322,9 @@ class TestFit(unittest.TestCase):
         else:
             transformed_data = scaler_x.transform(in_data)
 
-        test_oomd = model.calc_oomd(transformed_data, metric=metric)
+        test_oomd = model.calc_oomd(
+            input_array=transformed_data, metric=metric
+        )
         test_val = rmse(test_oomd, known_oomd)
         lin_val = nan_conc_coeff(test_oomd, known_oomd)
 
@@ -350,11 +356,11 @@ class TestFit(unittest.TestCase):
             transformed_data_y = scaler_y.transform(self.Y)
 
         model_low = NipalsPLS(n_components=1)
-        model_low.fit(transformed_data_x, transformed_data_y)
+        model_low.fit(input_x=transformed_data_x, input_y=transformed_data_y)
 
         # Update to new amount of components
         num_lvs = model.n_components
-        model_low.set_components(num_lvs)
+        model_low.set_components(n_component=num_lvs)
 
         # compare X scores
         if self.name == "Yes NaN, PLST RandomGen":
@@ -371,8 +377,8 @@ class TestFit(unittest.TestCase):
             self.assertGreater(lin_val, 1 - 1e-5, msg=f"linConc = {lin_val}")
 
         # compare X loadings
-        test_val = rmse(model.loadings_X, model_low.loadings_X)
-        lin_val = nan_conc_coeff(model.loadings_X, model_low.loadings_X)
+        test_val = rmse(model.loadings_x, model_low.loadings_x)
+        lin_val = nan_conc_coeff(model.loadings_x, model_low.loadings_x)
         with self.subTest():
             # overall rmse is low
             self.assertLess(test_val, 1e-3, msg=f"rmse = {test_val}")
[…]
