### 2025-05-23 9a6ac10
Use input naming conventions from baseclasses
