+++ src/open_nipals/nipalsPCA.py
@@ -19,13 +19,14 @@ from __future__ import (
     annotations,
 )  # needed so we can return NipalsPCA class in our type hints
 import numpy as np
+from sklearn.decomposition._base import _BasePCA
 import warnings
 from scipy.stats import f as F_dist
 from open_nipals.utils import _nan_mult
 from typing import Optional
 
 
-class NipalsPCA:
+class NipalsPCA(_BasePCA):
     """The custom-built class to use PCA using the NIPALS algorithm, i.e.,
     the same algorithm used in SIMCA.
 
@@ -265,14 +266,12 @@ class NipalsPCA:
 
         return self  # So methods can be chained
 
-    def transform(
-        self, input_array: np.ndarray, method: str = "naive"
-    ) -> np.ndarray:
+    def transform(self, X: np.ndarray, method: str = "naive") -> np.ndarray:
         """This function takes an input array and projects it based
         on a fitted model.
 
         Args:
-            input_array (np.ndarray): The nxm input array in the original
+            X (np.ndarray): The nxm input array in the original
                 feature space to be projected.
             method (str, optional): The method to use for the projection.
                 See reference listed in module docstring.
@@ -293,8 +292,8 @@ class NipalsPCA:
                 "Model has not yet been fit. Consider using fit_transform."
             )
 
-        # Extract shape of input_array
-        n, _ = input_array.shape
+        # Extract shape of X
+        n, _ = X.shape
 
         # Extract variables for simplicity
         num_lvs = self.n_components
@@ -302,7 +301,7 @@ class NipalsPCA:
 
         scores = np.zeros((n, num_lvs))
 
-        nan_mask = np.isnan(input_array)
+        nan_mask = np.isnan(X)
         nan_flag = np.any(nan_mask)
 
         if (not nan_flag) or (method == "naive"):
@@ -311,20 +310,18 @@ class NipalsPCA:
             for ind_lv in range(num_lvs):
                 if nan_flag:
                     scores[:, [ind_lv]] = _nan_mult(
-                        input_array,
+                        X,
                         loadings[:, [ind_lv]],
                         nan_mask,
                         use_denom=True,
                     )
                 else:
-                    scores[:, [ind_lv]] = input_array @ loadings[:, [ind_lv]]
+                    scores[:, [ind_lv]] = X @ loadings[:, [ind_lv]]
                     scores[:, [ind_lv]] = scores[:, [ind_lv]] / (
                         loadings[:, [ind_lv]].T @ loadings[:, [ind_lv]]
                     )
                 # deflate input data
-                input_array = (
-                    input_array - scores[:, [ind_lv]] @ loadings[:, [ind_lv]].T
-                )
+                X = X - scores[:, [ind_lv]] @ loadings[:, [ind_lv]].T
 
         elif nan_flag and (method == "projection"):
             # uses approach described in section 4 of McGregor paper
@@ -336,8 +333,7 @@ class NipalsPCA:
                     @ loadings[not_null, :num_lvs]
                 )  # denominator in eq. 9
                 nom = (
-                    loadings[not_null, :num_lvs].T
-                    @ input_array[row, not_null].T
+                    loadings[not_null, :num_lvs].T @ X[row, not_null].T
                 )  # nominator in eq. 9
                 scores[row, :] = (denom @ nom).T
 
@@ -372,23 +368,19 @@ class NipalsPCA:
                     S22 = P[not_null, :] @ theta @ P[not_null, :].T
 
                     # compute an estimate for the missing data
-                    z_hash = (
-                        S12 @ np.linalg.inv(S22) @ input_array[row, not_null].T
-                    )
-                    input_array[row, is_null] = z_hash
+                    z_hash = S12 @ np.linalg.inv(S22) @ X[row, not_null].T
+                    X[row, is_null] = z_hash
 
[…]
