### 2025-05-23 6a631b9
Update example code in README, corrected typo in merge
