+++ .github/workflows/python-app.yml
@@ -23,6 +23,9 @@ jobs:
       uses: actions/setup-python@v3
       with:
         python-version: "3.10"
+    - name: Check data files
+      if: ${{ hashFiles('data/') != '' }}
+      run: print ${{ hashFiles('data/') }}
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
@@ -33,7 +36,7 @@ jobs:
         # stop the build if there are Python syntax errors or undefined names
         flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
-        flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
+        flake8 . --count --exit-zero --max-complexity=10 --max-line-length=79 --statistics
     - name: Test with pytest
       run: |
         pytest
