+++ .github/workflows/python-app.yml
@@ -23,9 +23,8 @@ jobs:
       uses: actions/setup-python@v3
       with:
         python-version: "3.10"
-    - name: Check data files
-      if: ${{ hashFiles('data/') != '' }}
-      run: hashFiles('data/')
+    - name: XDTesting Print Directory Structure
+      uses: FiorelaCiroku/XDTesting-Print-Directory-Structure@v1.0.2
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
