+++ .github/workflows/python-app.yml
@@ -19,12 +19,12 @@ jobs:
 
     steps:
     - uses: actions/checkout@v4
-    - name: Set up Python 3.10
-      uses: actions/setup-python@v3
+    - name: Set up Python 3.13
+      uses: actions/setup-python@v5
       with:
-        python-version: "3.10"
-    - name: XDTesting Print Directory Structure
-      uses: FiorelaCiroku/XDTesting-Print-Directory-Structure@v1.0.2
+        python-version: "3.13"
+    # - name: XDTesting Print Directory Structure
+    #   uses: FiorelaCiroku/XDTesting-Print-Directory-Structure@v1.0.2
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
