+++ tox.ini
@@ -10,6 +10,13 @@ envlist =
 
 isolated_build = True
 
+[gh-actions]
+python = 
+	3.9: py39
+	3.10: py310
+	3.11: py311
+	3.12: py312
+	3.13: py313
 
 [testenv]
 description = Invoke pytest to run automated tests
