+++ .github/workflows/python-app.yml
@@ -34,5 +34,8 @@ jobs:
         flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=79 --statistics
+    - name: Check test file is there
+      run: |
+        if [-f /home/runner/work/open_nipals/open_nipals/data/SIMCA_ScaledNanDat_Loadings.xlsx ]; then echo "file is there"; fi
     - name: Test with tox
       run: tox
