+++ .github/workflows/python-app.yml
@@ -34,11 +34,17 @@ jobs:
         flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=79 --statistics
-    - name: Check test file is there
+    - name: Check relative file paths
       run: |
         ls -a data/
         if [ -f data/SIMCA_ScaledNanDat_Loadings.xlsx ]; then echo "xlsx file is there"; else echo "xlsx file absent"; fi
         if [ -f data/nanData.csv ]; then echo "csv file is there"; else echo "csv file absent"; fi
         if [ -f tests/test_nipalsPCA.py ]; then echo "test is there"; else echo "test absent"; fi
+    - name: Check absolute file paths
+      run: |
+        ls -a /home/runner/work/open_nipals/open_nipals/data/
+        if [ -f /home/runner/work/open_nipals/open_nipals/data/SIMCA_ScaledNanDat_Loadings.xlsx ]; then echo "xlsx file is there"; else echo "xlsx file absent"; fi
+        if [ -f /home/runner/work/open_nipals/open_nipals/data/nanData.csv ]; then echo "csv file is there"; else echo "csv file absent"; fi
+        if [ -f /home/runner/work/open_nipals/open_nipals/tests/test_nipalsPCA.py ]; then echo "test is there"; else echo "test absent"; fi
     # - name: Test with tox
     #   run: tox
