+++ .github/workflows/python-app.yml
@@ -37,14 +37,9 @@ jobs:
     - name: Check relative file paths
       run: |
         ls -a data/
-        if [ -f data/SIMCA_ScaledNanDat_Loadings.xlsx ]; then echo "xlsx file is there"; else echo "xlsx file absent"; fi
+        if [ -f data/SIMCA_ScaledNanDat_Loadings.xlsx ]; then echo "underscore xlsx file is there"; else echo "underscore xlsx file absent"; fi
+        if [ -f data/PCATestData.xlsx ]; then echo "no underscore xlsx file is there"; else echo "no underscore xlsx file absent"; fi
         if [ -f data/nanData.csv ]; then echo "csv file is there"; else echo "csv file absent"; fi
         if [ -f tests/test_nipalsPCA.py ]; then echo "test is there"; else echo "test absent"; fi
-    - name: Check absolute file paths
-      run: |
-        ls -a /home/runner/work/open_nipals/open_nipals/data/
-        if [ -f /home/runner/work/open_nipals/open_nipals/data/SIMCA_ScaledNanDat_Loadings.xlsx ]; then echo "xlsx file is there"; else echo "xlsx file absent"; fi
-        if [ -f /home/runner/work/open_nipals/open_nipals/data/nanData.csv ]; then echo "csv file is there"; else echo "csv file absent"; fi
-        if [ -f /home/runner/work/open_nipals/open_nipals/tests/test_nipalsPCA.py ]; then echo "test is there"; else echo "test absent"; fi
     # - name: Test with tox
     #   run: tox
