### 2025-05-23 9413572
See if raw strings for filenames are useful
