### 2025-05-26 adbafd8
next try
- remove the raw strings for input files
- only one python version tested by tox to save resources for tryout
- try commenting out df_simca_loads to see if it's that file specifically
