+++ .github/workflows/python-app.yml
@@ -14,7 +14,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-      python-version: ['3.13']
+        python-version: ['3.13']
         # python-version: ['3.9', '3.10', '3.11', '3.12', '3.13']
 
     steps:
