### 2025-05-26 dc2b321
Is it the usecols argument of read_excel?
