### 2025-05-26 45c89fd
Update test_nipalsPCA.py
next try...
