+++ tests/test_nipalsPCA.py
@@ -22,9 +22,10 @@ input_array = pca_dat.to_numpy()
 df_simca_loads = pd.read_excel(
     path.joinpath("SIMCA_ScaledFullDat_Loadings.xlsx"),
     engine="openpyxl",
-    # usecols=[1, 2],
-)  # First column is garbage index
-simca_loads = df_simca_loads.iloc[:, 1:2].to_numpy()
+)
+simca_loads = df_simca_loads.to_numpy()[
+    :, 1:2
+]  # First column is garbage index
 
 df_simca_sample_dat = pd.read_excel(
     path.joinpath("SIMCA_ScaledFullDat_Scores_T2Range_DMODXAbs.xlsx"),
@@ -45,9 +46,10 @@ input_array_nan = nan_dat.to_numpy()
 df_simca_loads = pd.read_excel(
     path.joinpath("SIMCA_ScaledNanDat_Loadings.xlsx"),
     engine="openpyxl",
-    # usecols=[1, 2],
-)  # First column is garbage index
-simca_loads_nan = df_simca_loads.iloc[:, 1:2].to_numpy()
+)
+simca_loads_nan = df_simca_loads.to_numpy()[
+    :, 1:2
+]  # First column is garbage index
 
 df_simca_sample_dat = pd.read_excel(
     path.joinpath("SIMCA_ScaledNanDat_Scores_T2Range_DMODXAbs.xlsx"),
