### 2025-05-26 2d1d445
Update test_nipalsPCA.py
Can we somehow delete the cache by commenting out in one push and then put it back in?
