### 2025-05-26 dfb036a
Update test_nipalsPCA.py
... and the last one
