### 2025-05-26 0d55297
Update python-app.yml
remove commented out code
