+++ .github/workflows/python-app.yml
@@ -22,8 +22,6 @@ jobs:
       uses: actions/setup-python@v5
       with:
         python-version: ${{ matrix.python-version }}
-    # - name: XDTesting Print Directory Structure
-    #   uses: FiorelaCiroku/XDTesting-Print-Directory-Structure@v1.0.2
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
@@ -34,12 +32,5 @@ jobs:
         flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=79 --statistics
-    # - name: Check relative file paths
-    #   run: |
-    #     ls -a data/
-    #     if [ -f data/SIMCA_ScaledNanDat_Loadings.xlsx ]; then echo "underscore xlsx file is there"; else echo "underscore xlsx file absent"; fi
-    #     if [ -f data/PCATestData.xlsx ]; then echo "no underscore xlsx file is there"; else echo "no underscore xlsx file absent"; fi
-    #     if [ -f data/nanData.csv ]; then echo "csv file is there"; else echo "csv file absent"; fi
-    #     if [ -f tests/test_nipalsPCA.py ]; then echo "test is there"; else echo "test absent"; fi
     - name: Test with tox
       run: tox
