### 2025-05-26 b2fe28f
Update test_nipalsPCA.py
Let's see if it works this time
