+++ tests/test_nipalsPCA.py
@@ -43,18 +43,18 @@ nan_dat = pd.read_excel(
 )
 input_array_nan = nan_dat.to_numpy()
 
-# df_simca_loads = pd.read_excel(
-#     path.joinpath("SIMCA_ScaledNanDat_Loadings.xlsx"),
-#     engine="openpyxl",
-# )
+df_simca_loads = pd.read_excel(
+    path.joinpath("SIMCA_ScaledNanDat_Loadings.xlsx"),
+    engine="openpyxl",
+)
 simca_loads_nan = df_simca_loads.to_numpy()[
     :, 1:2
 ]  # First column is garbage index
 
-# df_simca_sample_dat = pd.read_excel(
-#     path.joinpath("SIMCA_ScaledNanDat_Scores_T2Range_DMODXAbs.xlsx"),
-#     engine="openpyxl",
-# )
+df_simca_sample_dat = pd.read_excel(
+    path.joinpath("SIMCA_ScaledNanDat_Scores_T2Range_DMODXAbs.xlsx"),
+    engine="openpyxl",
+)
 simca_scores_nan = df_simca_sample_dat.to_numpy()[
     :, 1:3
 ]  # First column is garbage index
