### 2025-05-26 8136add
Update test_nipalsPCA.py
Fix typing
