+++ .github/workflows/python-app.yml
@@ -19,9 +19,6 @@ jobs:
 
     steps:
     - uses: actions/checkout@v4
-      with:
-        path: '.'
-        clean: 'false'
     - name: Set up Python ${{ matrix.python-version }}
       uses: actions/setup-python@v5
       with:
