### 2025-05-26 5c3f78e
Update python-app.yml
Run for all python versions
