### 2025-05-26 edf76c9
Update python-app.yml (#5)
Update `test_nipalsPCA.py` and `python-app.yml` such that github actions runs `flake8` and `tox`
---------

Co-authored-by: nielsschlusser <>
