### 2025-05-27 fcb087c
Update CONTRIBUTING.md
- `tox` is run automatically for all branches attempting to merge into the `main`-branch
- removed some small typos
