### 2025-05-27 947abf8
Hotfix/nonpos args (#4)
- base classes `_BasePCA` and `_PLS` invoked, arguments are adapted to accommodate for that
- made method arguments explicit in tests
- `tox` is run automatically via github actions for all branches attempting to merge into the `main`-branch -> adapt CONTRIBUTING.md
- updated README.md accordingly
---------

Co-authored-by: nielsschlusser <>
