### 2025-07-17 #8598 @editorialbot
[REVIEW]: open_nipals: An sklearn-compatible python package for NIPALS dimensional reduction
**Submitting author:** @nielsschlusser (<a href="http://orcid.org/0000-0002-3534-2153">Niels Schlusser</a>)
**Repository:** https://github.com/johnsonandjohnson/open_nipals
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.0.0
