### 2025-08-22 d408282
Make NipalsPLS interact with set_components() (#8)
Limited the matrices to the number of currently used components, while more may be actually fitted.

Co-authored-by: nielsschlusser <>
