### 2025-08-22 9cdb04b
Implemented explained_variance_ratio_ in NipalsPCA and NipalsPLS
- implement explained_variance_ratio_ for both, NipalsPCA and NipalsPLS
- implemented corresponding unit tests
- re-organized imports to comply with pylint policy
- currently still empty slice warning in NipalsPLS l. 799
