### 2025-09-20 #8598 @maximtrp
@nielsschlusser
I suggest deploying your package to [PyPI](https://pypi.org/) and hosting its documentation on [Read the Docs](https://readthedocs.org/). Cloning a repo to install a Python package isn’t standard practice, and nobody is going to clone a repo just to read the docs. These two steps are basically must-haves for usability and adoption. Both can be done using GitHub Action pipelines.
