### 2025-09-20 johnsonandjohnson/open_nipals#10 @rsenne
JOSS REVIEW - DOCS
Hi all. This is where I will host my review for documentation related things here. For the documentation I have a single review for right now: I can't seem to find any documentation? Do you have docs anywhere? If so can you put them in an easily accessible location or make it more apparent on the repo so I can read through them?
