### 2025-09-20 johnsonandjohnson/open_nipals#11 @rsenne
JOSS REVIEW - Paper
# Review: *open\_nipals* — an sklearn-compatible NIPALS package

## Summary (what it claims)

The paper presents an sklearn-style implementation of NIPALS for PCA/PLS with a focus on missing data (via Nelson’s Single Component Projection), scikit-learn pipeline compatibility, diagnostics (IMD/T^2 and OOMD using DModX/Q residuals), and a function to draw limits on the IMD–OOMD plane.

---

## Major points

### 1) Why NIPALS? Please teach the ML audience (briefly)

I’m an ML reviewer but unfamiliar with NIPALS. The manuscript contrasts NIPALS with SVD and highlights missing-data handling via SCP, but it doesn’t yet tell me **when** I should choose NIPALS over common workflows (simple/iterative impute → SVD PCA/PLS; PPCA/FA with EM; matrix completion + PCA).
Please add a short primer: what NIPALS is optimizing, how the iteration works at a high level, and practical trade-offs (speed, accuracy, robustness, convergence behavior) in regimes like tall/wide data with NaNs—ideally anchored in your manufacturing-with-missingness use case.

### 2) Convergence, defaults, and complexity

Because NIPALS is iterative, document stopping criteria, default tolerance and max iterations, recommended settings by data shape/sparsity, and computational/memory complexity vs. SVD. Also describe failure modes (non-convergence, oscillation) and how users should diagnose/mitigate them.

### 3) Validation: head-to-head baselines an sklearn user expects

To justify a dedicated package, show comparisons (or written descriptions) against:

* (A) mean/iterative impute → sklearn PCA/PLS,
* (B) PPCA/FA (EM) for missing entries,
* (C) matrix completion (e.g., soft-impute/ALS) → PCA,
* (D) PCR vs. PLS for regression.

Report explained variance or prediction error and wall-clock time across missingness rates and aspect ratios. This will substantiate the motivation beyond feasibility.

### 4) Figure 1

I'm admittedly a bit confused what I should be taking away from Figure 1. I can see that all data points fall inmside the 95% CI but the range is quite large, and generally feels a smidge underwhelming. Can you make this plot more convincing in a way that really shows your method is working in a really interesting way?

---

## Reproducibility and installation

Snippets are helpful, but please provide:

* A minimal, end-to-end example (dataset + notebook/script) that reproduces your figure(s) and the IMD/OOMD limits.
* Explicit installation instructions (pip/conda), supported Python versions, and dependency details.
* How randomness (if any) is controlled for reproducibility.

Link these from the “Availability” section and reference the exact tagged release used for the paper.

---

## IMD/OOMD math and assumptions

Include explicit formulas (with references) for Hotelling’s T^2, DModX, Q residuals, and the derivation/assumptions used by your limit calculation for the IMD–OOMD plane (e.g., F-distribution assumptions). Clarify any differences between PCA and PLS diagnostics in your implementation.

---

## Scope: standalone package vs. upstreaming

Your strongest selling point is sklearn compatibility (fit/transform, pipelines), which could argue for upstreaming into scikit-learn or a general decomposition library. Please add a short “Why not upstream?” paragraph (e.g., governance/scope constraints, domain-specific diagnostics, algorithm maturity, numerical concerns). If separation is essential, clarify unique scope (chemometrics diagnostics, ArrangeData semantics, domain plots) that wouldn’t fit upstream.

---

## API guarantees (sklearn parity)

Spell out which estimator APIs/tags you implement (`get_params`, `set_params`, `n_features_in_`, `feature_names_in_`, `inverse_transform` behavior with NaNs, and how `set_components` interacts with fitted state). This will help users understand drop-in compatibility and edge cases.

---

## Comparisons and guidance (help the user choose)

* **NIPALS (+SCP) vs. PPCA/FA (EM):**
  Briefly position assumptions and pros/cons (e.g., uncertainty quantification in PPCA, convergence behavior, accuracy under MCAR/MAR missingness). Provide practical guidance on when each is preferable.

* **PCR (PCA + linear regression) vs. PLS:**
  Since you ship both PCA and PLS, add guidance on when to pick PCR vs. PLS (bias/variance trade-offs, correlated predictors, small-n/large-p, multi-response Y). Confirm that predict/centering/scaling semantics match scikit-learn.

* **Modern low-rank + missing-data baselines:**
  Briefly position against common matrix-completion approaches (ALS/soft-impute) that many ML users reach for before specialized PCA variants. Even a qualitative comparison helps non-chemometrics readers.

---

## Minor/editorial

* **Fix typos and formatting:**
  “Principal Component Analysis” (not “Principle”); “transform” (not “tranform”); remove duplicated figure caption text (e.g., “Figure 1: Figure 1: …”); fix any broken references such as “(see (?))”.

* **ArrangeData contract:**
  Define what “correct ordering and quantity of input columns” means (schema learned at fit? handling of missing/extra columns? warnings vs. errors). Link to docstrings.

* **References and availability:**
  Add DOIs/URLs consistently, include the repository URL and the exact version/tag used in the paper body (not just in references), and include a CI badge and a short Testing & CI paragraph (coverage, what’s tested, how to run).

---

## Nice to have

* A small benchmark table (accuracy + runtime vs. %NaNs) comparing NIPALS(+SCP) to impute→SVD PCA/PLS, PPCA/FA, and soft-impute→PCA.
* Cross-validation helpers for choosing `n_components` (scree/explained-variance tooling suitable for NIPALS).
* Numerical stability notes (deflation/normalization choices; behavior with collinear variables).

---

**Bottom line**
The package looks useful and the sklearn-style API is welcome. To meet JOSS’s expectations on need and validation, I’d like to see: (i) a short NIPALS primer for the general ML reader, (ii) concrete baselines against common alternatives, (iii) clear convergence/complexity details, and (iv) tighter reproducibility and API documentation. I also would like to see a better argued statement of need: _I am not familiar with nipals as a method, convince me why I need it_. I think one of the easiest ways to argue this is to (i) compare against more well known methods and (ii) explain why this project merits a standalone package
