### 2025-11-18 abd19af
Update publish-to-test-pypi.yml
add `verbose: true` to upload to testpypi
