+++ .github/workflows/publish-to-test-pypi.yml
@@ -56,11 +56,9 @@ jobs:
     needs:
     - build
     runs-on: ubuntu-latest
-
     environment:
       name: testpypi
       url: https://test.pypi.org/p/open_nipals
-
     permissions:
       id-token: write  # IMPORTANT: mandatory for trusted publishing
 
@@ -74,3 +72,4 @@ jobs:
       uses: pypa/gh-action-pypi-publish@release/v1
       with:
         repository-url: https://test.pypi.org/legacy/
+        verbose: true
