### 2025-11-18 b643dec
Make compliant version numbers
made changes according to https://stackoverflow.com/questions/79433709/how-to-avoid-invalid-version-numbers-when-publishing-to-testpypi-in-github-actio
