+++ .github/workflows/publish-to-test-pypi.yml
@@ -71,5 +71,6 @@ jobs:
     - name: Publish distribution 📦 to TestPyPI
       uses: pypa/gh-action-pypi-publish@release/v1
       with:
+        skip-existing: true
         repository-url: https://test.pypi.org/legacy/
         verbose: true
