+++ pyproject.toml
@@ -4,8 +4,8 @@ requires = ["setuptools>=46.1.0", "setuptools_scm[toml]>=5", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [tool.setuptools_scm]
-# See configuration details in https://github.com/pypa/setuptools_scm
-version_scheme = "no-guess-dev"
+write_to = {{…}}
+local_scheme = "no-local-version"
 
 [tool.ruff]
 line-length = 79
