### 2025-11-18 e36bef6
Update pyproject.toml
