+++ pyproject.toml
@@ -4,7 +4,6 @@ requires = ["setuptools>=46.1.0", "setuptools_scm[toml]>=5", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [tool.setuptools_scm]
-write_to = {{…}}
 local_scheme = "no-local-version"
 
 [tool.ruff]
