### 2025-11-18 f2e5069
Remove verbose flag
