### 2025-11-19 0c79bd6
build docs using sphinx_rtd_theme
- set up documentation with sphinx-quickstart
- built documentation with tox -e docs
- view docs with python -m http.serve from docs/ dir and visiting localhost:8000 in browser
