+++ tox.ini
@@ -34,57 +34,57 @@ deps =
     sklearn15: scikit-learn >= 1.5, <1.7
 
 
-# [testenv:{build,clean}]
-# description =
-#     build: Build the package in isolation according to PEP517, see https://github.com/pypa/build
-#     clean: Remove old distribution files and temporary build artifacts (./build and ./dist)
-# # https://setuptools.pypa.io/en/stable/build_meta.html#how-to-use-it
-# skip_install = True
-# changedir = {toxinidir}
-# deps =
-#     build: build[virtualenv]
-# passenv =
-#     SETUPTOOLS_*
-# commands =
-#     clean: python -c 'import shutil; [shutil.rmtree(p, True) for p in ("build", "dist", "docs/_build")]'
-#     clean: python -c 'import pathlib, shutil; [shutil.rmtree(p, True) for p in pathlib.Path("src").glob("*.egg-info")]'
-#     build: python -m build {posargs}
+[testenv:{build,clean}]
+description =
+    build: Build the package in isolation according to PEP517, see https://github.com/pypa/build
+    clean: Remove old distribution files and temporary build artifacts (./build and ./dist)
+# https://setuptools.pypa.io/en/stable/build_meta.html#how-to-use-it
+skip_install = True
+changedir = {toxinidir}
+deps =
+    build: build[virtualenv]
+passenv =
+    SETUPTOOLS_*
+commands =
+    clean: python -c 'import shutil; [shutil.rmtree(p, True) for p in ("build", "dist", "docs/_build")]'
+    clean: python -c 'import pathlib, shutil; [shutil.rmtree(p, True) for p in pathlib.Path("src").glob("*.egg-info")]'
+    build: python -m build {posargs}
 
 
-# [testenv:{docs,doctests,linkcheck}]
-# description =
-#     docs: Invoke sphinx-build to build the docs
-#     doctests: Invoke sphinx-build to run doctests
-#     linkcheck: Check for broken links in the documentation
-# passenv =
-#     SETUPTOOLS_*
-# setenv =
-#     DOCSDIR = {toxinidir}/docs
-#     BUILDDIR = {toxinidir}/docs/_build
-#     docs: BUILD = html
-#     doctests: BUILD = doctest
-#     linkcheck: BUILD = linkcheck
-# deps =
-#     -r {toxinidir}/docs/requirements.txt
-#     # ^  requirements.txt shared with Read The Docs
-# commands =
-#     sphinx-build --color -b {env:BUILD} -d "{env:BUILDDIR}/doctrees" "{env:DOCSDIR}" "{env:BUILDDIR}/{env:BUILD}" {posargs}
+[testenv:{docs,doctests,linkcheck}]
+description =
+    docs: Invoke sphinx-build to build the docs
+    doctests: Invoke sphinx-build to run doctests
+    linkcheck: Check for broken links in the documentation
+passenv =
+    SETUPTOOLS_*
+setenv =
+    DOCSDIR = {toxinidir}/docs
+    BUILDDIR = {toxinidir}/docs/_build
+    docs: BUILD = html
+    doctests: BUILD = doctest
+    linkcheck: BUILD = linkcheck
+deps =
+    -r {toxinidir}/docs/requirements.txt
+    # ^  requirements.txt shared with Read The Docs
+commands =
+    sphinx-build --color -b {env:BUILD} -d "{env:BUILDDIR}/doctrees" "{env:DOCSDIR}" "{env:BUILDDIR}/{env:BUILD}" {posargs}
 
 
-# [testenv:publish]
-# description =
-#     Publish the package you have been developing to a package index server.
-#     By default, it uses testpypi. If you really want to publish your package
[…]
