### 2025-11-19 bf83f05
Included autodoc in conf.py extensions, rebuilt
