+++ tox.ini
@@ -7,8 +7,10 @@ minversion = 3.24
 envlist = 
     py{39,310,311}-sklearn{10,15}
     py{312,313}-sklearn{15}
+    py314-sklearn17
 
 isolated_build = True
+
 
 [gh-actions]
 python = 
@@ -17,6 +19,8 @@ python =
 	3.11: py311
 	3.12: py312
 	3.13: py313
+    3.14: py314
+
 
 [testenv]
 description = Invoke pytest to run automated tests
@@ -32,6 +36,7 @@ commands =
 deps = 
     sklearn10: scikit-learn >= 1.0, < 1.5
     sklearn15: scikit-learn >= 1.5, <1.7
+    sklearn17: scikit-learn >= 1.7
 
 
 [testenv:{build,clean}]
