+++ setup.cfg
@@ -39,17 +39,17 @@ package_dir =
     =src
 
 # Require a min/specific Python version (comma-separated conditions)
-python_requires = >=3.9,<3.14
+python_requires = >=3.9,<3.15
 
 # Add here dependencies of your project (line-separated), e.g. requests>=2.2,<3.0.
 # Version specifiers like >=2.2,<3.0 avoid problems due to API changes in
 # new major versions. This works if the required packages follow Semantic Versioning.
 # For more information, check out https://semver.org/.
 install_requires =
-    numpy >=1.21, <2.3
-    pandas >=2.0, <2.3
-    scipy >=1.5, <1.15
-    scikit-learn >=1.2, <1.7
+    numpy >=1.21, <2.4
+    pandas >=2.0, <2.4
+    scipy >=1.5, <1.17
+    scikit-learn >=1.2, <1.8
 
 # scipy >= 1.14 conflicts with python 3.9, so for recent scipy use python >= 3.10 
 # pandas <= 2.2 conflicts with numpy >= 1.26.4

[…]
