### 2025-12-09 a2471b7
Docs created with sphinx
- create docs with sphinx
- adapted docstrings in src/open_nipals/ folder
