+++ docs/conf.py
@@ -19,7 +19,11 @@ release = "1.0.1"
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
 
-extensions = ["sphinx.ext.autodoc"]
+# enable automated creation of docs with autodoc
+# conversion of numpy/autodocstring style to
+# rst readable style with napoleon
+# rendering markdown with myst
+extensions = ["sphinx.ext.autodoc", "sphinx.ext.napoleon", "myst_parser"]
 
 templates_path = ["_templates"]
 exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
