+++ docs/readme.rst
@@ -1,0 +1,5 @@
+README
+===========
+
+.. include:: ../README.md
+   :parser: myst_parser.sphinx_

[…]
