### 2025-12-10 a3c79ea
Update documentation
- include .readthedocs.yaml config file
- change README.md pypi installation paragraph
- recompile the docs
