+++ .readthedocs.yaml
@@ -9,19 +9,11 @@ version: 2
 build:
    os: ubuntu-22.04
    tools:
-     python: "3.10"
-     # You can also specify other tool versions:
-     # nodejs: "16"
-     # rust: "1.55"
-     # golang: "1.17"
+     python: "3.13"
 
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
    configuration: docs/conf.py
-
-# If using Sphinx, optionally build your docs in additional formats such as PDF
-# formats:
-#    - pdf
 
 # Optionally declare the Python requirements required to build your docs
 python:
