### 2025-12-10 5b900e7
Update nipalsPLS.py
Address rsenne's Small patch candidate 1.
