### 2025-12-10 5056978
Guard convergence division
- guard convergence division against zero norm in both modules, adressing rsenne's small patch candidates 2.
- restructure calc_imd in nipalsPCA.py in line with rsenne's small patch candidate suggestion 1
