### 2025-12-10 66c7eb9
Update nipalsPCA.py
Use pseudo-inverse in nipalsPCA.py projection bit in transform to address rsenne's small patch candidates 3.
