### 2025-12-10 77f8ee3
Change print to warning if iteration stops prematurely
- change print to warning statement in case iterations are terminated due to reaching maximum number of iterations in both modules, addressing rsenne's small patch candidate 4.
- use fstrings in the prints for verbose option instead of  .format(...)
