### 2025-12-10 2c83318
Relax default values + use np concatenation for performance
- use np.concatenate  instead of np.append for performance gains, addressing rsenne's small patch candidates 6
- relax the default relative tolerance criteria to 1e-6, addressing rsenne's comment on performance
