### 2025-12-10 c634884
Update nipalsPLS.py
Correct bug in NIpalsPLS.calc_imd when using scores
