### 2025-12-10 3e00eb8
Update nipalsPCA.py
Missed another inv->pinv
