### 2025-12-10 b2204d0
Update nipalsPLS.py
unify exponential numbers format
