### 2025-12-10 3bcf21b
Inherit from different classes
- change base classes for NipalsPCA from _BasePCA (private, unstable) to BaseEstimator, TransformerMixin (public)
- change base classes for NipalsPLS from _PLS (private, unstable) to BaseEstimator, TransformerMixin, RegressorMixin (public)
- re-organize inputs to comply with pylint input organization hierarchy (python native, thrid party, first party)
