+++ src/open_nipals/nipalsPCA.py
@@ -17,16 +17,16 @@ revised 2024: Niels Schlusser
 from __future__ import (
     annotations,
 )  # needed so we can return NipalsPCA class in our type hints
-import numpy as np
-from sklearn.decomposition._base import _BasePCA
-from sklearn.exceptions import NotFittedError
 import warnings
+from typing import Optional
+import numpy as np
+from sklearn.base import BaseEstimator, TransformerMixin
+from sklearn.exceptions import NotFittedError
 from scipy.stats import f as F_dist
 from open_nipals.utils import _nan_mult
-from typing import Optional
 
 
-class NipalsPCA(_BasePCA):
+class NipalsPCA(BaseEstimator, TransformerMixin):
     """The custom-built class to use PCA using the NIPALS algorithm, i.e.,
     the same algorithm used in SIMCA.
 
