### 2025-12-10 f857a67
Implement comments in both modules, re-create docs
- fit new components in NipalsPCA transform conditionalMean with set_components, adddressing rsenne's major issue 5
- updated and streamlined documentation where necessary
- streamlined NipalsPLS calc_oomd, adapted to NipalsPCA counterpart
- re-create documentation
