### 2025-12-11 cd47631
Working on full covariance matrix for Hotelling's T2
- Ledoit Wolf does still not work due to dimensionality issue
- updated docs
