### 2025-12-12 ecc0749
New options to calc_imd for both modules, updated docs
- two more options for calc_imd in both modules: full using the full covariance matrix, and ledoit_wolf for using the full covariance matrix using the Ledoit Wolf shrinkage for performance
- update the docs
