+++ src/open_nipals/nipalsPLS.py
@@ -63,7 +63,7 @@ class NipalsPLS(BaseEstimator, TransformerMixin, RegressorMixin):
         The fitted y scores of the PLS model.
     regression_matrix : np.ndarray
         The regression matrix of the PLS model.
-    fitted_components : int:
+    fitted_components : int
         The number of current LVs in the model (0 if not fitted yet.)
 
     Methods:
@@ -593,9 +593,11 @@ class NipalsPLS(BaseEstimator, TransformerMixin, RegressorMixin):
                 Must be one of set {'HotellingT2'}.
                 Defaults to 'HotellingT2'.
             covariance (str, optional): Method to compute covariance. Valid
-                options are {'diag', 'full'}. Defaults to 'diag'
-                (quick version). 'full' uses the entire covariance matrix
-                computed by Ledoit-Wolf shrinkage.
+                options are {'diag', 'full', 'ledoit_wolf'}.
+                Defaults to 'diag' (quick version).
+                'full' uses the entire covariance matrix computed by numpy.
+                'ledoit_wolf' uses the full covariance matrix computed
+                by Ledoit-Wolf shrinkage.
 
         Raises:
             NotFittedError: If model has not been fit.
@@ -654,21 +656,32 @@ class NipalsPLS(BaseEstimator, TransformerMixin, RegressorMixin):
                         (scores - fit_means) ** 2 / fit_vars, axis=1
                     ).reshape(-1, 1)
                 elif covariance == "full":
-                    # Ledoit Wolf shrinkage
-                    # TODO: this still needs work
-                    # dimensionalities doe not work out yet
+                    # use full covariance matrix
+                    out_imd = np.diagonal(
+                        (scores - fit_means)
+                        @ np.linalg.pinv(
+                            np.cov(
+                                self.fit_scores_x[:, :num_lvs_fit].T, ddof=1
+                            )
+                        )
+                        @ (scores - fit_means).T
+                    ).reshape(-1, 1)
+                elif covariance == "ledoit_wolf":
+                    # compute full covariance matrix
+                    # with Ledoit-Wolf shrinkage
                     lw_obj = LedoitWolf(
                         assume_centered=self.mean_centered
                     ).fit(self.fit_scores_x[:, :num_lvs_fit])
-                    out_imd = (
+                    out_imd = np.diagonal(
                         (scores - fit_means)
                         @ np.linalg.pinv(lw_obj.covariance_)
-                        @ (scores - fit_means)
-                    )
+                        @ (scores - fit_means).T
+                    ).reshape(-1, 1)
                 else:
                     raise NotImplementedError(
                         f"Covariance method {covariance} not implemented."
-                        + "Needs to be in {'diag', 'full'}."
+                        + "Possible methods are "
+                        + "{'diag', 'full', 'ledoit_wolf'}."
                     )
 
         else:
