+++ docs/authors.rst
@@ -1,5 +1,10 @@
 Contributors' list
-===========
+==================
+
+.. image:: jnjim_logo.svg
+
+open\_nipals was originally developed at 
+Johnson & Johnson Innovative Medicine.
 
 .. include:: ../AUTHORS.md
    :parser: myst_parser.sphinx_
