+++ docs/modules.rst
@@ -1,7 +1,0 @@
-open_nipals
-===========
-
-.. toctree::
-   :maxdepth: 4
-
-   open_nipals
