### 2025-12-12 543dcd7
Use snake_case for method options in nipalsPCA, update docs
