+++ docs/contributing.rst
@@ -1,5 +1,5 @@
 How can I contribute?
-===========
+=====================
 
 .. include:: ../CONTRIBUTING.md
    :parser: myst_parser.sphinx_
