### 2025-12-16 2830042
Draft eplained_variance_ and explained_variance_ratio_ methods with unit tests
