### 2026-01-05 25f160b
Streamline with sklearn, fix unit tests
- split get_explained_variance and get_Explained_variance_ratio methods from properties
- both methods now return numpy arrays
- fixed the unit tests
