### 2026-01-06 johnsonandjohnson/open_nipals#12 @nielsschlusser
Major Issues:
1. We have changed the base classes from the private ones to the ones you suggested. We have also exposed some of the attributes you suggested. As our package is designed to operate after standardizing the data with `sklearn.StandardScaler`, we refrained from exposing `mean_ `.
2. For precisely that reason, we decided not to include an automatic mean centering step into our package, also, because in case anyone _really_ knows what they are doing, we would like to grant them the opportunity to use the package on non-mean-centered data. We have pointed out in the documentation that we _highly_ recommend to mean_center the data.
3. Thank you very much for pointing out the bug. We have fixed it.
4. According to your suggestion, we have replaced the direct matrix inverses by pseudo-inverses.
5. Thank you for pointing out the re-fit of the entire model in the `conditional_mean` case of `transform`. This was indeed a very costly and suboptimal implementation. We have replaced it by a double use of the `set_components()` method, see [code](https://github.com/johnsonandjohnson/open_nipals/blob/7d66eb590e8667b2c6f8462f7618dd249c046701/src/open_nipals/nipalsPCA.py#L379-L387).
6. We have expanded the `calc_imd()` function by a `covariance` option, that can be either of `{'diag', 'full', 'ledoit_wolf'}`.
7. Thank you for pointing out the lack of consistency between `calc_oomd()` for PCA and PLS. Unfortunately, this inconsistency seems to originate from SIMCA itself. In order to maintain consistency with SIMCA, we left it as is.
