### 2026-01-06 johnsonandjohnson/open_nipals#12 @nielsschlusser
Performance:
Default tolerance values of `1e-8/1e-10` are indeed too strict. We have set a less rigid default value of `1e-6`.
