### 2026-01-06 #8598 @nielsschlusser
Hi @maximtrp, @maximtrp, @yuanqing-wang,

Happy New Year to all of you!

Thank you very much for the helpful comments. Following the above suggestions, we have configured `github-actions` to automatically deploy the package to `PyPI` once the changes are on the `main` branch. Same holds true for the docs, which have been updated.

Code changes addressing @rsenne's comments, as well as further comments, are collected under [this PR](https://github.com/johnsonandjohnson/open_nipals/pull/13).
