### 2026-01-07 ba046c3
Improve normalization for NipalsPLS
- nipalsPCA.py: remove use_Denom=True statements, as this is the default value anyway
- nipalsPLS.py: follow the recipe of K. Dunn's book for normalization
- update the documentation
