### 2026-01-08 e186fbe
Update nipalsPLS.py
Correct normalization
