### 2026-01-09 814521c
Tighter limits on PLS unit tests
- implement tighter limits on PLS unit tests
- remove regression vector computation from set_Components unit test
