+++ tests/test_nipalsPLS_RandomGen.py
@@ -234,10 +234,15 @@ class TestFit(unittest.TestCase):
         """Compare fitted scores to scores from package (T)"""
         test_val = rmse(self.T, self.model[0].fit_scores_x)
         lin_val = nan_conc_coeff(self.T, self.model[0].fit_scores_x)
+
+        # tolerances per dataset
         if self.name == "Yes NaN, PLST RandomGen":
-            err_lim = 1e-1
+            err_lim = 5e-2
+        elif self.name == "No NaN, PLST RandomGen":
+            err_lim = 5e-3
         else:
-            err_lim = 1e-2
+            err_lim = 5e-5
+
         with self.subTest():
             # overall rmse is low
             self.assertLess(test_val, err_lim, msg=f"rmse = {test_val}")
@@ -268,15 +273,24 @@ class TestFit(unittest.TestCase):
             self.assertLess(test_val, 1e-9, msg=f"rmse = {test_val}")
         with self.subTest():
             # correlation is very high
-            self.assertGreater(lin_val, 1 - 1e-6, msg=f"linConc = {lin_val}")
+            self.assertGreater(lin_val, 1 - 1e-9, msg=f"linConc = {lin_val}")
 
     def test_fit_loadings(self):
         """Compare loadings to loadings from package (P)"""
         test_val = rmse(self.P, self.model[0].loadings_x)
         lin_val = nan_conc_coeff(self.P, self.model[0].loadings_x)
+
+        # tolerances per dataset
+        if self.name == "Yes NaN, PLST RandomGen":
+            err_lim = 5e-5
+        elif self.name == "No NaN, PLST RandomGen":
+            err_lim = 5e-6
+        else:
+            err_lim = 5e-8
+
         with self.subTest():
             # overall rmse is low
-            self.assertLess(test_val, 1e-3, msg=f"rmse = {test_val}")
+            self.assertLess(test_val, err_lim, msg=f"rmse = {test_val}")
         with self.subTest():
             # correlation is very high
             self.assertGreater(lin_val, 1 - 1e-5, msg=f"linConc = {lin_val}")
@@ -292,10 +306,15 @@ class TestFit(unittest.TestCase):
         )  # reverse standardscaling
         test_val = rmse(self.yhat, py_y_vals)
         lin_val = nan_conc_coeff(self.yhat, py_y_vals)
+
+        # tolerances per dataset
         if self.name == "Yes NaN, PLST RandomGen":
-            err_lim = 1e-1
+            err_lim = 5e-2
+        elif self.name == "No NaN, PLST RandomGen":
+            err_lim = 5e-3
         else:
-            err_lim = 1e-2
+            err_lim = 5e-4
+
         with self.subTest():
             # overall rmse is low
             self.assertLess(test_val, err_lim, msg=f"rmse = {test_val}")
@@ -312,9 +331,16 @@ class TestFit(unittest.TestCase):
         py_reg_vects = py_reg_vects / np.linalg.norm(py_reg_vects)  # Normalize
         test_val = rmse(b, py_reg_vects)
         lin_val = nan_conc_coeff(b, py_reg_vects)
+
+        # tolerances per dataset
+        if self.name == "Yes NaN, PLST RandomGen":
+            err_lim = 5e-5
+        else:
+            err_lim = 5e-7
+
         with self.subTest():
             # overall rmse is low
-            self.assertLess(test_val, 1e-3, msg=f"rmse = {test_val}")
+            self.assertLess(test_val, err_lim, msg=f"rmse = {test_val}")
         with self.subTest():
             # correlation is very high
             self.assertGreater(lin_val, 1 - 1e-5, msg=f"linConc = {lin_val}")
[…]
