### 2026-04-19 #8598 @maximtrp
@nielsschlusser Hi! Could you please point me to the documentation? I thought you would include a link to publicly available docs (for example, at https://readthedocs.org as I suggested).
