### 2026-04-19 #8598 @nielsschlusser
Hey @maximtrp ,

Sure, it's [here](https://open-nipals.readthedocs.io/en/latest/). Further ressources can be found in the PRs @rsenne opened:
- [docs](https://github.com/johnsonandjohnson/open_nipals/issues/10)
- [code](https://github.com/johnsonandjohnson/open_nipals/issues/12)
- [paper](https://github.com/johnsonandjohnson/open_nipals/issues/11)

The PyPI deployment of the package can be found under [this link](https://pypi.org/project/open-nipals/).
