### 2026-05-01 #8598 @maximtrp
I've finished my review. Please add a link to your ReadTheDocs page to README.md and to the repo info (so that it's easy to find). Now I don't see it anywhere.

Also, some issues noticed by Claude Code (pay attention to 1, 4, 5):
1. Broken DOI link (Citation section of README.md) — it should be https://doi.org/10.5281/zenodo.18375840
2. Badges — none exist at all (optional, IMO, but good to have). A professional repo top should have at minimum:
  - Build/CI status (GitHub Actions)
  - PyPI version (pip install open-nipals)
  - Supported Python versions
  - License
  - ReadTheDocs status
  - Zenodo DOI
3. Suggested test matrix:
```
[tox]
envlist =
    py{39,310,311}-sklearn{12,15}
    py{312,313}-sklearn{15,17}
    py314-sklearn{15,17}

[testenv]
deps =
    sklearn12: scikit-learn >=1.2, <1.5
    sklearn15: scikit-learn >=1.5, <1.7
    sklearn17: scikit-learn >=1.7, <1.8
```
4. setup.cfg issues found while reading
- `url` points to a specific PR (/pull/13) instead of the repo root.
- Documentation project URL points to https://about.readthedocs.com/ (the ReadTheDocs marketing homepage) instead of the actual package docs.
- `Development Status :: 4 - Beta` — a published, stable, v2.0 package should be `5 - Production/Stable`.
- Only two PyPI classifiers — missing Python version classifiers (`Programming Language :: Python :: 3.9`, etc.) and topic classifiers (`Topic :: Scientific/Engineering :: Chemistry` etc.) that help with discoverability.
5. Suggested requirements:
```
install_requires =
    numpy >=1.24, <3
    pandas >=2.0, <3
    scipy >=1.9, <2
    scikit-learn >=1.2, <1.8
```
And please check package compatibility with the latest pandas version (3.0.1). If compatible, you may ease off the upper version bound to <4.

Thank you guys and good luck!
